move path mapping out of ExecutionEngine and into DataConnector for a better encapsulation
