+++ great_expectations/datasource/data_connector/configured_asset_dbfs_data_connector.py
@@ -62,14 +62,8 @@ class ConfiguredAssetDBFSDataConnector(ConfiguredAssetFilesystemDataConnector):
     def _get_full_file_path_for_asset(
         self, path: str, asset: Optional[Asset] = None
     ) -> str:
-
         full_path = super()._get_full_file_path_for_asset(path=path, asset=asset)
-
         template_arguments: dict = {
             "path": full_path,
         }
-
-        return self.execution_engine.resolve_data_reference(
-            data_connector_name=self.__class__.__name__,
-            template_arguments=template_arguments,
-        )
+        return self.resolve_data_reference(template_arguments=template_arguments)
