+++ great_expectations/datasource/data_connector/configured_asset_gcs_data_connector.py
@@ -160,7 +160,4 @@ class ConfiguredAssetGCSDataConnector(ConfiguredAssetFilePathDataConnector):
             "bucket_or_name": self._bucket_or_name,
             "path": path,
         }
-        return self.execution_engine.resolve_data_reference(
-            data_connector_name=self.__class__.__name__,
-            template_arguments=template_arguments,
-        )
+        return self.resolve_data_reference(template_arguments=template_arguments)
