+++ great_expectations/datasource/data_connector/configured_asset_s3_data_connector.py
@@ -141,7 +141,4 @@ class ConfiguredAssetS3DataConnector(ConfiguredAssetFilePathDataConnector):
             "bucket": self._bucket,
             "path": path,
         }
-        return self.execution_engine.resolve_data_reference(
-            data_connector_name=self.__class__.__name__,
-            template_arguments=template_arguments,
-        )
+        return self.resolve_data_reference(template_arguments=template_arguments)
