+++ great_expectations/datasource/data_connector/inferred_asset_dbfs_data_connector.py
@@ -55,16 +55,10 @@ class InferredAssetDBFSDataConnector(InferredAssetFilesystemDataConnector):
     def _get_full_file_path(
         self, path: str, data_asset_name: Optional[str] = None
     ) -> str:
-
         full_path = super()._get_full_file_path(
             path=path, data_asset_name=data_asset_name
         )
-
         template_arguments: dict = {
             "path": full_path,
         }
-
-        return self.execution_engine.resolve_data_reference(
-            data_connector_name=self.__class__.__name__,
-            template_arguments=template_arguments,
-        )
+        return self.resolve_data_reference(template_arguments=template_arguments)
