+++ great_expectations/datasource/data_connector/inferred_asset_s3_data_connector.py
@@ -152,10 +152,7 @@ class InferredAssetS3DataConnector(InferredAssetFilePathDataConnector):
             "bucket": self._bucket,
             "path": path,
         }
-        return self.execution_engine.resolve_data_reference(
-            data_connector_name=self.__class__.__name__,
-            template_arguments=template_arguments,
-        )
+        return self.resolve_data_reference(template_arguments=template_arguments)
 
 
 def _check_valid_s3_path(
